+++ utils/bitfield/src/rleplus/mod.rs
@@ -155,3 +155,87 @@ impl BitField {
         writer.finish()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::{
+        super::{bitfield, ranges_from_bits},
+        BitField, BitWriter,
+    };
+
+    use rand::{Rng, SeedableRng};
+    use rand_xorshift::XorShiftRng;
+
+    #[test]
+    fn test() {
+        for (bits, expected) in vec![
+            (vec![], bitfield![]),
+            (
+                vec![
+                    0, 0, // version
+                    1, // starts with 1
+                    0, 1, // fits into 4 bits
+                    0, 0, 0, 1, // 8 - 1
+                ],
+                bitfield![1, 1, 1, 1, 1, 1, 1, 1],
+            ),
+            (
+                vec![
+                    0, 0, // version
+                    1, // starts with 1
+                    0, 1, // fits into 4 bits
+                    0, 0, 1, 0, // 4 - 1
+                    1, // 1 - 0
+                    0, 1, // fits into 4 bits
+                    1, 1, 0, 0, // 3 - 1
+                ],
+                bitfield![1, 1, 1, 1, 0, 1, 1, 1],
+            ),
+            (
+                vec![
+                    0, 0, // version
+                    1, // starts with 1
+                    0, 0, // does not fit into 4 bits
+                    1, 0, 0, 1, 1, 0, 0, 0, // 25 - 1
+                ],
+                bitfield![
+                    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
+                ],
+            ),
+            // when a length of 0 is encountered, the rest of the encoded bits should be ignored
+            (
+                vec![
+                    0, 0, // version
+                    1, // starts with 1
+                    1, // 1 - 1
+                    0, 1, // fits into 4 bits
+                    0, 0, 0, 0, // 0 - 0
+                    1, // 1 - 1
+                ],
+                bitfield![1],
+            ),
+        ] {
+            let mut writer = BitWriter::new();
+            for bit in bits {
+                writer.write(bit, 1);
+            }
+            let bf = BitField::from_bytes(&writer.finish()).unwrap();
+            assert_eq!(bf, expected);
+        }
+    }
+
+    #[test]
+    fn roundtrip() {
+        let mut rng = XorShiftRng::seed_from_u64(1);
+
+        for _i in 0..1000 {
+            let len: usize = rng.gen_range(0, 1000);
+            let bits: Vec<_> = (0..len).filter(|_| rng.gen::<bool>()).collect();
+
+            let ranges: Vec<_> = ranges_from_bits(bits.clone()).collect();
+            let bf = BitField::from_ranges(ranges_from_bits(bits));
+
+            assert_eq!(bf.ranges().collect::<Vec<_>>(), ranges);
+        }
+    }
+}
